Move the pressurecurve calculation into a separate file
There's nothing driver-specific about this and it allows us to test it
separately or write tools to test it.
