+++ src/wcmFilter.c
@@ -22,17 +22,7 @@
 #include <math.h>
 #include "xf86Wacom.h"
 #include "wcmFilter.h"
-
-/*****************************************************************************
- * Static functions
- ****************************************************************************/
-
-static void filterCurveToLine(int* pCurve, int nMax, double x0, double y0,
-		double x1, double y1, double x2, double y2,
-		double x3, double y3);
-static int filterOnLine(double x0, double y0, double x1, double y1,
-		double a, double b);
-static void filterLine(int* pCurve, int nMax, int x0, int y0, int x1, int y1);
+#include "wcmPressureCurve.h"
 
 
 /*****************************************************************************
@@ -101,128 +91,6 @@ void wcmResetSampleCounter(const WacomChannelPtr pChannel)
 }
 
 
-static void filterNearestPoint(double x0, double y0, double x1, double y1,
-		double a, double b, double* x, double* y)
-{
-	double vx, vy, wx, wy, d1, d2, c;
-
-	wx = a - x0; wy = b - y0;
-	vx = x1 - x0; vy = y1 - y0;
-
-	d1 = vx * wx + vy * wy;
-	if (d1 <= 0)
-	{
-		*x = x0;
-		*y = y0;
-	}
-	else
-	{
-		d2 = vx * vx + vy * vy;
-		if (d1 >= d2)
-		{
-			*x = x1;
-			*y = y1;
-		}
-		else
-		{
-			c = d1 / d2;
-			*x = x0 + c * vx;
-			*y = y0 + c * vy;
-		}
-	}
-}
-
-static int filterOnLine(double x0, double y0, double x1, double y1,
-		double a, double b)
-{
-        double x, y, d;
-	filterNearestPoint(x0,y0,x1,y1,a,b,&x,&y);
-	d = (x-a)*(x-a) + (y-b)*(y-b);
-	return d < 0.00001; /* within 100th of a point (1E-2 squared) */
-}
-
-static void filterCurveToLine(int* pCurve, int nMax, double x0, double y0,
-		double x1, double y1, double x2, double y2,
-		double x3, double y3)
-{
-	double x01,y01,x32,y32,xm,ym;
-	double c1,d1,c2,d2,e,f;
-
-	/* check if control points are on line */
-	if (filterOnLine(x0,y0,x3,y3,x1,y1) && filterOnLine(x0,y0,x3,y3,x2,y2))
-	{
-		filterLine(pCurve,nMax,
-			(int)(x0*nMax),(int)(y0*nMax),
-			(int)(x3*nMax),(int)(y3*nMax));
-		return;
-	}
-
-	/* calculate midpoints */
-	x01 = (x0 + x1) / 2; y01 = (y0 + y1) / 2;
-	x32 = (x3 + x2) / 2; y32 = (y3 + y2) / 2;
-
-	/* calc split point */
-	xm = (x1 + x2) / 2; ym = (y1 + y2) / 2;
-
-	/* calc control points and midpoint */
-	c1 = (x01 + xm) / 2; d1 = (y01 + ym) / 2;
-	c2 = (x32 + xm) / 2; d2 = (y32 + ym) / 2;
-	e = (c1 + c2) / 2; f = (d1 + d2) / 2;
-
-	/* do each side */
-	filterCurveToLine(pCurve,nMax,x0,y0,x01,y01,c1,d1,e,f);
-	filterCurveToLine(pCurve,nMax,e,f,c2,d2,x32,y32,x3,y3);
-}
-
-static void filterLine(int* pCurve, int nMax, int x0, int y0, int x1, int y1)
-{
-	int dx, dy, ax, ay, sx, sy, x, y, d;
-
-	/* sanity check */
-	if ((x0 < 0) || (y0 < 0) || (x1 < 0) || (y1 < 0) ||
-		(x0 > nMax) || (y0 > nMax) || (x1 > nMax) || (y1 > nMax))
-		return;
-
-	dx = x1 - x0; ax = abs(dx) * 2; sx = (dx>0) ? 1 : -1;
-	dy = y1 - y0; ay = abs(dy) * 2; sy = (dy>0) ? 1 : -1;
-	x = x0; y = y0;
-
-	/* x dominant */
-	if (ax > ay)
-	{
-		d = ay - ax / 2;
-		while (1)
-		{
-			pCurve[x] = y;
-			if (x == x1) break;
-			if (d >= 0)
-			{
-				y += sy;
-				d -= ax;
-			}
-			x += sx;
-			d += ay;
-		}
-	}
-
-	/* y dominant */
-	else
-	{
-		d = ax - ay / 2;
-		while (1)
-		{
-			pCurve[x] = y;
-			if (y == y1) break;
-			if (d >= 0)
-			{
-				x += sx;
-				d -= ay;
-			}
-			y += sy;
-			d += ax;
-		}
-	}
-}
 static void storeRawSample(WacomCommonPtr common, WacomChannelPtr pChannel,
 			   WacomDeviceStatePtr ds)
 {
